==> analysis/01_wrangle_code.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%cd "Documents/GitHub/pmx_gasoline_theft_prediction"

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

dict_replacers_string = {"Delegación":"","Deleg. ":"","Del. ":""} 

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].replace(dict_replacers_string)

for k, v in dict_replacers_string.items():
    df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace(k, v)               




df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)




df_join_pipe_code['codigo inegi'] = df_join_pipe_code['codigo inegi'].astype(str).str.zfill(7)

df_join_pipe_code['codigo inegi'] = [x[:-2] for x in df_join_pipe_code['codigo inegi']]



len(df_join_pipe_code['codigo inegi'])

df_join_pipe_code_var = df_join_pipe_code[["year", "tomas", "codigo inegi", "estado", "municipio"]]

df_codes_mx['codigo inegi'] = df_codes_mx['codigo inegi'].astype(str).str.zfill(7)

df_codes_mx['codigo inegi'] = [x[:-2] for x in df_codes_mx['codigo inegi']]

df_join_pipe_code_final = pd.merge(df_join_pipe_code_var,
                             df_codes_mx,
                             on='codigo inegi',
                             how='outer',
                             indicator=True)

  %notebook analysis/01_Wrangle_data_PMX_code.ipynb

df_join_pipe_code.to_excel('data/df_join_pipe_code.xlsx', index = False)

df_join_pipe_code_final.to_excel('data/df_join_pipe_code_final.xlsx', index = False)

df_join_pipe_code.dtypes #final


df_join_pipe_code_final_2019 = df_join_pipe_code_final[(df_join_pipe_code_final['year'] == 2019)]

df_join_pipe_code_final_2019.to_excel('data/df_join_pipe_code_final_2019.xlsx', index = False)

df_join_pipe_code_final_2019_all = pd.read_excel("data/df_join_pipe_code_final_2019_full.xlsx")


df_pipe_code_final_2019_var = df_join_pipe_code_final_2019_all[["year", "tomas", "codigo inegi"]]

df_pipe_code_final_2019_var

df_pipe_code_final_2019_var['codigo inegi'] = df_pipe_code_final_2019_var['codigo inegi'].astype(str).str.zfill(5)

df_pipe_code_final_2019_var_final = df_pipe_code_final_2019_var.rename(columns = {'codigo inegi' : 'CVEGEO',
                                              'tomas' : 'Num_thefts'})
df_pipe_code_final_2019_var_final



df_pipe_code_final_2019_var_final.to_excel('data/df_pipe_code_final_2019_var_final.xlsx', index = False)

df_pipe_code_final_2019_var_final.to_parquet('data/df_pipe_code_final_2019_var_final.parquet')
