==> analysis/02_wrangle_pipelines.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import pandas as pd
import glob
import os

import pyarrow as pa
import pyarrow.parquet as pq


%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"


path = 'data-raw' # use your path
filename = glob.glob(path + "/*.csv")

all_files = glob.glob(os.path.join(path , "*.csv"))

df_from_each_file = (pd.read_csv(f) for f in all_files)


concatenated_df   = pd.concat(df_from_each_file, ignore_index=True)
    
#frame = pd.concat(li, axis=0, ignore_index=True)

df_INEGI = pd.read_csv("data-raw/INEGI.csv")

df_five_year_pipeline = pd.read_csv("data-raw/five_year_pipeline.csv")

df_installations = pd.read_csv("data-raw/installations.csv")

df_integ_sistran_pp = pd.read_csv("data-raw/integrated_sistrangas_pipeline.csv")

df_mix_pipeline = pd.read_csv("data-raw/mix_pipeline.csv")

df_not_integ_sistran_pp = pd.read_csv("data-raw/not_integrated_sistrangas_pipeline.csv")

df_oil_pipelines = pd.read_csv("data-raw/oil_pipelines.csv")

df_pipeline_transport_system = pd.read_csv("data-raw/pipeline_transport_system.csv")

df_project_five_years = pd.read_csv("data-raw/project_five_years.csv")

df_project_gas_pipeline = pd.read_csv("data-raw/project_gas_pipeline.csv")

df_SISTRANGAS_pipeline = pd.read_csv("data-raw/SISTRANGAS_pipeline.csv")

df_storage_terminal = pd.read_csv("data-raw/storage_terminal.csv")






dict_inegi = {"COUNT": "pp_inegi","numero":"type_inegi",
                       "rela_suel":"soil_inegi"} 

df_INEGI_w = df_INEGI[["CVEGEO", "numero", "rela_suel", "COUNT"]].rename(columns = dict_inegi)

df_INEGI_w['CVEGEO'] = df_INEGI_w['CVEGEO'].astype(str).str.zfill(5)



dict_fy_pipeline = {"COUNT": "pp_5ypp","CAP_MMPCD":"cap_5ypp",
                       "INV_MMUSD":"inv_5ypp", "LONGITUD_K": "long_5ypp"} 

df_five_year_pipeline_w = df_five_year_pipeline[["CVEGEO", "CAP_MMPCD", "INV_MMUSD", "LONGITUD_K", "COUNT"]].rename(columns = dict_fy_pipeline)

df_five_year_pipeline_w['CVEGEO'] = df_five_year_pipeline_w['CVEGEO'].astype(str).str.zfill(5)



df_installations

dict_installations = {"COUNT": "pp_instal"} 

df_installations_w = df_installations[["CVEGEO", "COUNT"]].rename(columns = dict_installations)

df_installations_w['CVEGEO'] = df_installations_w['CVEGEO'].astype(str).str.zfill(5)



df_integ_sistran_pp


dict_integ_sistran_pp = {"COUNT": "pp_integ_sistran_pp",
                        "longitud_k": "long_integ_sistran_pp",
                        "tipo":"type_integ_sistran_pp"} 

df_integ_sistran_pp_w = df_integ_sistran_pp[["CVEGEO", "COUNT", "longitud_k", "tipo"]].rename(columns = dict_integ_sistran_pp)

df_integ_sistran_pp_w['CVEGEO'] = df_integ_sistran_pp_w['CVEGEO'].astype(str).str.zfill(5)



df_mix_pipeline

dict_df_mix_pipeline = {"COUNT": "pp_mix_pp",
                        "longitud_k": "long_mix_pp",
                        "tipo_de_tu":"type_tube_mix_pp",
                       "tipo_de_hi":"type_gas_mix_pp"} 

df_mix_pipeline_w = df_mix_pipeline[["CVEGEO", "COUNT", "longitud_k", "tipo_de_tu", "tipo_de_hi"]].rename(columns = dict_df_mix_pipeline)

df_mix_pipeline_w['CVEGEO'] = df_mix_pipeline_w['CVEGEO'].astype(str).str.zfill(5)




df_not_integ_sistran_pp

dict_not_integ_sistran_pp = {"COUNT": "pp_n_int_sist",
                        "longitud_k": "long_n_int_sist",
                        "tipo": "type_n_int_sist",
                        "capacidad_":"cap_n_int_sist",
                        "inversin_m":"inv_n_int_sist"} 

df_not_integ_sistran_pp_w = df_not_integ_sistran_pp[["CVEGEO", "COUNT", "longitud_k", "tipo", "capacidad_", "inversin_m"]].rename(columns = dict_not_integ_sistran_pp)


df_not_integ_sistran_pp_w['CVEGEO'] = df_not_integ_sistran_pp_w['CVEGEO'].astype(str).str.zfill(5)



df_oil_pipelines


dict_oil_pipelines = {"COUNT": "pp_oil",
                        "longitud": "long_oil",
                        "servicio": "type_oil",
                        "capacidad_":"cap_oil",
                        "capacida_1":"cap_oil_2"} 

df_oil_pipelines_w = df_oil_pipelines[["CVEGEO", "COUNT", "longitud", "servicio", "capacidad_", "capacida_1"]].rename(columns = dict_oil_pipelines)


df_oil_pipelines_w['CVEGEO'] = df_oil_pipelines_w['CVEGEO'].astype(str).str.zfill(5)


df_pipeline_transport_system


dict_pipeline_transport_system = {"COUNT": "pp_pp_tran_sys","CAP_MMPCD":"cap_pp_tran_sys",
                       "INV_MMUSD":"inv_pp_tran_sys", "LONGITUD_K": "long_pp_tran_sys"} 

df_pipeline_transport_system_w = df_five_year_pipeline[["CVEGEO", "CAP_MMPCD", "INV_MMUSD", "LONGITUD_K", "COUNT"]].rename(columns = dict_pipeline_transport_system)



df_pipeline_transport_system_w['CVEGEO'] = df_pipeline_transport_system_w['CVEGEO'].astype(str).str.zfill(5)



df_project_five_years

dict_project_five_years = {"COUNT": "pp_proj_5y","CAP_MMPCD":"cap_proj_5y"} 

df_project_five_years_w = df_project_five_years[["CVEGEO", "CAP_MMPCD", "COUNT"]].rename(columns = dict_project_five_years)

df_project_five_years_w['CVEGEO'] = df_project_five_years_w['CVEGEO'].astype(str).str.zfill(5)



df_project_gas_pipeline

dict_project_gas_pipeline = {"COUNT": "pp_proj_gas_p","INV_MMUSD":"inv_proj_gas_p",
                            "LONGITUD_K": "long_proj_gas_p"} 

df_project_gas_pipeline_w = df_project_gas_pipeline[["CVEGEO", "INV_MMUSD", "COUNT", "LONGITUD_K"]].rename(columns = dict_project_gas_pipeline)


df_project_gas_pipeline_w['CVEGEO'] = df_project_gas_pipeline_w['CVEGEO'].astype(str).str.zfill(5)



df_SISTRANGAS_pipeline



dict_SISTRANGAS_pipeline = {"COUNT": "pp_SISTRANGAS_pipeline","CAP_MMPCD":"cap_SISTRANGAS_pipeline",
                       "INV_MMUSD":"inv_SISTRANGAS_pipeline", "LONGITUD_K": "long_SISTRANGAS_pipeline"} 

df_SISTRANGAS_pipeline_w = df_SISTRANGAS_pipeline[["CVEGEO", "CAP_MMPCD", "INV_MMUSD", "LONGITUD_K", "COUNT"]].rename(columns = dict_SISTRANGAS_pipeline)


df_SISTRANGAS_pipeline_w['CVEGEO'] = df_SISTRANGAS_pipeline_w['CVEGEO'].astype(str).str.zfill(5)





df_storage_terminal


dict_storage_terminal = {"COUNT": "pp_st_term",
                        "servicio": "type_st_term",
                        "capacidad_":"cap_st_term",
                        "capacida_1":"cap_st_term_1"} 

df_storage_terminal_w = df_oil_pipelines[["CVEGEO", "COUNT", "servicio", "capacidad_", "capacida_1"]].rename(columns = dict_storage_terminal)


df_storage_terminal_w['CVEGEO'] = df_storage_terminal_w['CVEGEO'].astype(str).str.zfill(5)






dfs = [df_INEGI_w, 
df_five_year_pipeline_w, 
df_installations_w,
df_integ_sistran_pp_w, 
df_mix_pipeline_w, 
df_not_integ_sistran_pp_w,
df_oil_pipelines_w,
df_pipeline_transport_system_w, 
df_project_five_years_w,
df_project_gas_pipeline_w, 
df_SISTRANGAS_pipeline_w, 
df_storage_terminal_w]



df_pipelines_mx = pd.concat(dfs, join='outer')

dfs = [df.set_index('id') for df in dfList]


from functools import reduce
df_pipelines_mx = reduce(lambda df1,df2: pd.merge(df1,df2,on='CVEGEO'), dfs)

conditions = [
    
    (df_pipelines_mx['pp_inegi'] >= 1),

    (df_pipelines_mx['pp_5ypp'] >= 1),

    (df_pipelines_mx['pp_integ_sistran_pp'] >= 1),

    (df_pipelines_mx['pp_instal'] >= 1),

    (df_pipelines_mx['pp_mix_pp'] >= 1),
    
    (df_pipelines_mx['pp_n_int_sist'] >= 1),

    (df_pipelines_mx['pp_oil'] >= 1),
    
    (df_pipelines_mx['pp_pp_tran_sys'] >= 1),

    (df_pipelines_mx['pp_proj_5y'] >= 1),

    (df_pipelines_mx['pp_proj_gas_p'] >= 1),
    
    (df_pipelines_mx['pp_SISTRANGAS_pipeline'] >= 1),

    (df_pipelines_mx['pp_st_term'] >= 1),

]

letters = [1] * 12

 

df_pipelines_mx['pipeline'] = np.select(conditions, letters)

sum(df_pipelines_mx['pipeline'])

column_list = list(df_pipelines_mx)




def keep(item):
    # Implement whatever criterion you need here
     return item in ('pp_inegi',
'pp_5ypp',
'pp_integ_sistran_pp', 
'pp_instal',
'pp_mix_pp',
'pp_n_int_sist',
'pp_oil',
'pp_pp_tran_sys',
'pp_proj_5y',
'pp_proj_gas_p',
"pp_SISTRANGAS_pipeline",
'pp_st_term')

filter(keep, column_list)

newlist = list(filter(keep, column_list))

df_pipelines_mx["sum_intersects"] = df_pipelines_mx[newlist].sum(axis=1)


df_pipelines_mx.to_parquet('df_pipelines_mx.parquet')
#table_mx = pa.Table.from_pandas(df_pipelines_mx)
#pd.write_table(table_mx, 'df_pipelines_mx.parquet')
#dataset_parquet = pd.read_parquet('dataset.parquet')

df_pipelines_mx.to_excel('data/df_pipelines_mx.xlsx', index = False)



df_PMX_08_15 = pd.read_excel("data-raw/PMXyearxy.xlsx")
