==> analysis/03_wrangle_places and_prices.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_prices_raw = pd.read_excel("data-raw/places_f.xlsx")
df_places_raw = pd.read_excel("data-raw/prices_pmx.xlsx")

df_prices = pd.read_excel("data-raw/prices_pmx.xlsx")


df_places = pd.read_csv("data-raw/places_gas.csv")

df_places['CVEGEO'] = df_places['CVEGEO'].astype(str).str.zfill(5)

df_places_w = df_places[["CVEGEO", "COUNT"]]

df_places_prices_raw = pd.merge(df_places_raw,
                             df_prices_raw,
                             on='place_id',
                             how='outer',
                             indicator=True)

df_places_prices_raw = df_places_prices_raw[(df_places_prices_raw['x'] != 0)]

df_places_prices_raw.to_excel('data/df_places_prices_raw.xlsx', index = False)

df_places_prices = pd.read_csv("data-raw/spatial_joint_gasoline_prices.csv")
df_places_prices['CVEGEO'] = df_places_prices['CVEGEO'].astype(str).str.zfill(5)

df_places_prices

#df_places_prices_group = df_places_prices.groupby("CVEGEO")["SELECTED"].sum()

df_places_prices_group = df_places_prices.groupby("CVEGEO").mean().reset_index()

df_places_prices_group = df_places_prices_group[["CVEGEO", "price"]]

df_places_prices_group

df_places_prices_final = pd.merge(df_places_w,
                             df_places_prices_group,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)



df_places_prices_final["price_15_19"] = df_places_prices_final["price"]*0.780871886#mean price 15-19
df_places_prices_final

df_places_prices_final["price"].mean()

#Prices 2019

df_places_prices_final["price_19"] = df_places_prices_final["price"]*0.883131673#mean price 15-19
df_places_prices_final

df_places_prices_final.to_excel('data/df_places_prices_final.xlsx', index = False)
df_places_prices_final.to_parquet('data/df_places_prices_final.parquet')
