==> analysis/04_wrangle_thefts.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"


df_PMX_SP_15_19 = pd.read_parquet("data/df_pipethefts_15_19.parquet").rename(columns = {'Mun_code' : 'CVEGEO'})

df_PMX_SP_19 = pd.read_excel("data/df_pipe_code_final_2019_var_final.xlsx")

df_PMX_08_15 = pd.read_excel("data-raw/PMXyearxy.xlsx")

df_PMX_08_15_w = df_PMX_08_15

df_PMX_SP_15_19
df_PMX_SP_19

df_PMX_08_15_w_f = df_PMX_08_15_w.drop_duplicates(subset = "CVEGEO")
df_PMX_08_15_w_final = df_PMX_08_15_w_f[["CVEGEO", "tomas_r"]]

df_PMX_08_15_w_final

df_PMX_SP_15_19_CC = pd.merge(df_PMX_SP_15_19,
                             df_PMX_08_15_w_final,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)

df_PMX_SP_15_19_CC['Num_thefts'] = df_PMX_SP_15_19_CC['Num_thefts'].fillna(0)

df_PMX_SP_15_19_CC['CVEGEO'] = df_PMX_SP_15_19_CC['CVEGEO'].dropna()



df_PMX_SP_15_19_CC_f = df_PMX_SP_15_19_CC[["CVEGEO", "Num_thefts"]]


df_PMX_SP_15_19_CC_f

df_PMX_SP_15_19_CC_f.to_excel('data/df_PMX_SP_15_19_thefts_pp.xlsx', index = False)
df_PMX_SP_15_19_CC_f.to_parquet('data/df_PMX_SP_15_19_thefts_pp.parquet')
