==> analysis/05_wrangle_pipelines_and_thefts.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_PMX_SP_15_19_CC_f = pd.read_parquet("data/df_PMX_SP_15_19_thefts_pp.parquet")

df_PMX_SP_19 = pd.read_excel("data/df_pipe_code_final_2019_var_final.xlsx")

df_pipelines_mx = pd.read_parquet("data/df_pipelines_mx.parquet")

df_PMX_SP_19['CVEGEO'] = df_PMX_SP_19['CVEGEO'].astype(str).str.zfill(5)


df_pipelines_mx['CVEGEO'] = df_pipelines_mx['CVEGEO'].astype(str)
df_PMX_SP_19['CVEGEO'] = df_PMX_SP_19['CVEGEO'].astype(str)
df_PMX_SP_15_19_CC_f['CVEGEO'] = df_PMX_SP_15_19_CC_f['CVEGEO'].astype(str)

df_PMX_SP_19

df_PMX_15_19_thefts_pp = pd.merge(df_pipelines_mx,
                             df_PMX_SP_15_19_CC_f,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)

df_PMX_19_thefts_pp = pd.merge(df_pipelines_mx,
                             df_PMX_SP_19,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)


df_PMX_15_19_thefts_pp

df_PMX_19_thefts_pp


df_PMX_15_19_thefts_pp.loc[df_PMX_15_19_thefts_pp["Num_thefts"] == 0, "pipeline"] = 1
df_PMX_19_thefts_pp.loc[df_PMX_19_thefts_pp["Num_thefts"] == 0, "pipeline"] = 1

df_PMX_19_thefts_pp

df_PMX_15_19_thefts_pp["pipeline"].sum()
df_PMX_19_thefts_pp["pipeline"].sum()


df_PMX_19_thefts_pp

df_PMX_15_19_thefts_pp.to_excel('data/df_PMX_15_19_thefts_pp_final.xlsx', index = False)
df_PMX_15_19_thefts_pp.to_parquet('data/df_PMX_15_19_thefts_pp_final.parquet')


df_PMX_19_thefts_pp.to_excel('data/df_PMX_19_thefts_pp_final.xlsx', index = False)
df_PMX_19_thefts_pp.to_parquet('data/df_PMX_19_thefts_pp_final.parquet')
