==> analysis/07.6_wrangle_rural.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_rural = pd.read_csv("data-raw/ageb_urbana_rural.csv")
df_rural_w = df_rural.drop(['CVEGEO', "CVE_LOC"], axis=1)
df_rural_w


df_rural_w['CVE_ENT'] = df_rural_w['CVE_ENT'].astype(str).str.zfill(2)

df_rural_w['CVE_MUN'] = df_rural_w['CVE_MUN'].astype(str).str.zfill(3)

df_rural_w['CVEGEO'] = df_rural_w['CVE_ENT'] + df_rural_w['CVE_MUN']

df_rural_w

df_rural_w.loc[df_rural_w["Ambito"] == "Urbana", "urban"] = 1

df_rural_w.loc[df_rural_w["Ambito"] == "Rural", "rural"] = 1


df_rural_w_group = df_rural_w.groupby(['CVEGEO'], as_index=0).sum()

df_rural_w_group['prop_rural'] = df_rural_w_group['rural'] / df_rural_w_group['urban']

df_rural_w_group

#df_cars_2019_f = df_cars_2019[["CVEGEO", "vehicles"]]

#df_cars_2019_f

df_rural_w_group.to_excel('data/df_rural_w_group.xlsx', index = False)
df_rural_w_group.to_parquet('data/df_rural_w_group.parquet')
