==> analysis/07.8_wrangle_density.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_density_km = pd.read_excel("data-raw/BASE-INFRAESTRUCTURA.xlsx", 'DATOS')
df_density_km_w = df_density_km.drop(['CARR-MT', "RED-MT", "AREAMT ", "NOM_MUN"], axis=1)



df_density_km_w

dict_var_density = {"CVEMUN": "CVEGEO"}


df_density_km_w = df_density_km_w.rename(columns = dict_var_density)

df_density_km_w['CVEGEO'] = df_density_km_w['CVEGEO'].astype(str).str.zfill(5)
df_density_km_w

df_density_km_w.to_excel('data/df_density_km.xlsx', index = False)
df_density_km_w.to_parquet('data/df_density_km.parquet')

pipe_thefts = pd.read_excel("data/df_pipethefts_15_19.xlsx")

dict_var_density = {"Mun_code": "CVEGEO"}


pipe_thefts = pipe_thefts.rename(columns = dict_var_density)

pipe_thefts['CVEGEO'] = pipe_thefts['CVEGEO'].astype(str).str.zfill(5)


pipe_thefts


pipe_thefts_final = pd.merge(df_density_km_w,
                             pipe_thefts,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)

pipe_thefts_final["Num_thefts"] = pipe_thefts_final["Num_thefts"].fillna(0)

pipe_thefts_final['CVEGEO'] = pipe_thefts_final['CVEGEO'].astype(str).str.zfill(5)


pipe_thefts_final.to_csv('data/df_pipe_CVEGEO.csv')

pipe_thefts_final
