==> analysis/07.55_wrangle_cars_mean.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_cars_2015 = pd.read_excel("data-raw/vmrc_anual_tr_cifra_2015.xlsx")
df_cars_2016 = pd.read_excel("data-raw/vmrc_anual_tr_cifra_2016.xlsx")
df_cars_2017 = pd.read_excel("data-raw/vmrc_anual_tr_cifra_2017.xlsx")
df_cars_2018 = pd.read_excel("data-raw/vmrc_anual_tr_cifra_2018.xlsx")
df_cars_2019 = pd.read_excel("data-raw/vmrc_anual_tr_cifra_2019.xlsx")





def add_variables(df):
    
    df['CVE_ENT'] = df['ID_ENTIDAD'].astype(str).str.zfill(2)
    
    df['CVE_MUN'] = df['ID_MUNICIPIO'].astype(str).str.zfill(3)
    
    df['CVEGEO'] = df['CVE_ENT'] + df['CVE_MUN']
    
    df['vehicles'] = df['AUTO_OFICIAL'] + df['AUTO_PUBLICO'] + df['AUTO_PARTICULAR'] + df['CAM_PAS_OFICIAL'] + df['CAM_PAS_PUBLICO'] + df['CAM_PAS_PARTICULAR'] + df['CYC_CARGA_OFICIAL'] + df_cars_2019['CYC_CARGA_PUBLICO'] + df_cars_2019['CYC_CARGA_PARTICULAR'] + df_cars_2019['MOTO_OFICIAL'] + df_cars_2019['MOTO_DE_ALQUILER'] + df_cars_2019['MOTO_PARTICULAR'] 
    
    df_f = df[["CVEGEO", "vehicles"]]
    
    return df_f


df_list = [df_cars_2015, df_cars_2016, df_cars_2017, df_cars_2018, df_cars_2019]

df_cars_2015, df_cars_2016, df_cars_2017, df_cars_2018, df_cars_2019 = [add_variables(df) for df in df_list]






dataframes = [df_cars_2015, df_cars_2016, df_cars_2017, df_cars_2018, df_cars_2019]




df_vehicles_final = reduce(lambda  left,right: pd.merge(left,right,on=['CVEGEO'],
                                            how = 'outer'), dataframes)#.fillna(0)

df_vehicles_final = df_vehicles_final.fillna(0)

df_vehicles_final

dict_vec = {"S_2015": "vehicles_x"} 

df_vehicles_final_w = df_vehicles_final.rename(columns = dict_vec)

df_vehicles_final_w.iloc[:, 1] + df_vehicles_final_w.iloc[:, 2]

df_vehicles_final["vehicles_15_19"] = df_vehicles_final_w.iloc[:, 1] + df_vehicles_final_w.iloc[:, 2] + df_vehicles_final_w.iloc[:, 3] + df_vehicles_final_w.iloc[:, 4] + df_vehicles_final_w.iloc[:, 5]/5

df_vehicles_final = df_vehicles_final[["CVEGEO", "vehicles_15_19"]]

df_vehicles_final

df_vehicles_final.to_excel('data/df_vehicles_final.xlsx', index = False)
df_vehicles_final.to_parquet('data/df_vehicles_final.parquet')

def add_variables(df):
    
df['CVE_ENT'] = df['ID_ENTIDAD'].astype(str).str.zfill(2)

df['CVE_MUN'] = df['ID_MUNICIPIO'].astype(str).str.zfill(3)

df['CVEGEO'] = df['CVE_ENT'] + df['CVE_MUN']

df['vehicles'] = df['AUTO_OFICIAL'] + df['AUTO_PUBLICO'] + df['AUTO_PARTICULAR'] + df['CAM_PAS_OFICIAL'] + df['CAM_PAS_PUBLICO'] + df['CAM_PAS_PARTICULAR'] + df['CYC_CARGA_OFICIAL'] + df_cars_2019['CYC_CARGA_PUBLICO'] + df_cars_2019['CYC_CARGA_PARTICULAR'] + df_cars_2019['MOTO_OFICIAL'] + df_cars_2019['MOTO_DE_ALQUILER'] + df_cars_2019['MOTO_PARTICULAR'] 

df_f = df[["CVEGEO", "vehicles"]]
    return df_f


df_list = [df_cars_2015, df_cars_2016, df_cars_2017, df_cars_2018, df_cars_2019]

df_cars_2015, df_cars_2016, df_cars_2017, df_cars_2018, df_cars_2019 = [add_variables(df) for df in df_list]






df_cars_2019['CVE_ENT'] = df_cars_2019['ID_ENTIDAD'].astype(str).str.zfill(2)

df_cars_2019['CVE_MUN'] = df_cars_2019['ID_MUNICIPIO'].astype(str).str.zfill(3)

df_cars_2019['CVEGEO'] = df_cars_2019['CVE_ENT'] + df_cars_2019['CVE_MUN']

df_cars_2019['vehicles'] = df_cars_2019['AUTO_OFICIAL'] + df_cars_2019['AUTO_PUBLICO'] + df_cars_2019['AUTO_PARTICULAR'] + df_cars_2019['CAM_PAS_OFICIAL'] + df_cars_2019['CAM_PAS_PUBLICO'] + df_cars_2019['CAM_PAS_PARTICULAR'] + df_cars_2019['CYC_CARGA_OFICIAL'] + df_cars_2019['CYC_CARGA_PUBLICO'] + df_cars_2019['CYC_CARGA_PARTICULAR'] + df_cars_2019['MOTO_OFICIAL'] + df_cars_2019['MOTO_DE_ALQUILER'] + df_cars_2019['MOTO_PARTICULAR'] 



df_cars_2019_f = df_cars_2019[["CVEGEO", "vehicles"]]

df_cars_2019_f

df_cars_2019_f.to_excel('data/df_cars_2019.xlsx', index = False)
df_cars_2019_f.to_parquet('data/df_cars_2019.parquet')

dict_SEDENA = {"MILITAR_MUERTO_NUM": "deaths_militar","MILITAR_HERIDO_NUM":"injur_militar",
                       "AGRESOR_MUERTO_NUM":"deaths_aggressor", "AGRESOR_HERIDO_NUM":"injur_aggressor",
               "AGRESOR_DETENIDO_NUM":"arrest_aggressor", "CLAVE_MUNICIPIO": "CVEGEO", "enfren": "confront"} 

df_SEDENA_08_19_w = df_SEDENA_08_19[["year", "CLAVE_MUNICIPIO", "MILITAR_HERIDO_NUM", "MILITAR_MUERTO_NUM", "AGRESOR_MUERTO_NUM", "AGRESOR_HERIDO_NUM", "AGRESOR_DETENIDO_NUM", "enfren"]].rename(columns = dict_SEDENA)

df_SEDENA_08_19_w

df_SEDENA_08_19_w_group = df_SEDENA_08_19_w.groupby(['year', 'CVEGEO'], as_index=0).sum()

df_SEDENA_08_19_w_group

df_SEDENA_group_15_19 = df_SEDENA_08_19_w_group[(df_SEDENA_08_19_w_group['year']>=2015)]

df_SEDENA_group_15_19

df_SEDENA_group_15_19["mortality_index"] = df_SEDENA_group_15_19["deaths_aggressor"]/df_SEDENA_group_15_19["injur_aggressor"]

df_SEDENA_group_15_19_w = df_SEDENA_group_15_19.groupby(['CVEGEO'], as_index=0).mean()

df_SEDENA_group_15_19_w_f = df_SEDENA_group_15_19_w[["CVEGEO", "confront"]]

df_SEDENA_group_15_19_w_f["confront"] = df_SEDENA_group_15_19_w["confront"].round()

df_SEDENA_group_15_19_w_f

df_SEDENA_group_15_19_w_f.to_excel('data/df_SEDENA_15_19.xlsx', index = False)
df_SEDENA_group_15_19_w_f.to_parquet('data/df_SEDENA_15_19.parquet')
