==> analysis/08_final_merge_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_homics_15_19 = pd.read_parquet("data/df_homics_15_19.parquet")
df_PMX_15_19_thefts_pp = pd.read_parquet('data/df_PMX_15_19_thefts_pp_final.parquet')
df_PMX_gas_place_price = pd.read_parquet('data/df_places_prices_final.parquet')
df_social_mean_2015_2019 = pd.read_excel('data/df_social_15_19_final_JR.xlsx')
df_SEDENA_15_19 = pd.read_parquet('data/df_SEDENA_15_19.parquet')
df_cvegeo = pd.read_excel("data-raw/codigo municipio inegi.xlsx")
df_cvegeo = df_cvegeo.rename(columns = {"codigo inegi":"CVEGEO"})

df_cvegeo['CVEGEO'] = df_cvegeo['CVEGEO'].astype(str).str.zfill(7).str.replace('.',"")
df_cvegeo['CVEGEO'] = df_cvegeo['CVEGEO'].str[:-1]
df_cvegeo_f = df_cvegeo[["CVEGEO", "NOM_ENT", "NOM_MUN"]]

df_social_mean_2015_2019['CVEGEO']=df_social_mean_2015_2019['CVEGEO'].astype(object)
df_social_mean_2015_2019['CVEGEO'] = df_social_mean_2015_2019['CVEGEO'].astype(str).str.zfill(5)

df_PMX_15_19_thefts_pp

df_cvegeo_f

df_homics_15_19.dtypes

df_PMX_15_19_thefts_pp.dtypes

df_PMX_15_19_thefts_pp



dataframes = [df_cvegeo_f,
              df_PMX_15_19_thefts_pp,
                    df_PMX_gas_place_price,
                    df_homics_15_19,
                    df_SEDENA_15_19,
                    df_social_mean_2015_2019]

df_PMX_thefts_15_19 = reduce(lambda  left,right: pd.merge(left,right,on=['CVEGEO'],
                                            how = 'outer'), dataframes)#.fillna(0)

df_PMX_thefts_15_19_pp = df_PMX_thefts_15_19[(df_PMX_thefts_15_19['pipeline']==1) & (df_PMX_thefts_15_19['CVEGEO'].str.len()>=5)]

df_homics_15_19

df_PMX_thefts_15_19_pp["homic_rate"] = df_PMX_thefts_15_19_pp["homic_15_19"]/df_PMX_thefts_15_19_pp["POB_TOT"]*100000


df_PMX_thefts_15_19_pp['confront']=df_PMX_thefts_15_19_pp['confront'].fillna(0)

df_PMX_thefts_15_19_pp_f = df_PMX_thefts_15_19_pp.drop(['_merge_x', '_merge_y', "pipeline"], axis=1)


df_PMX_thefts_15_19_pp_f['pipeline'] = np.select([(df_PMX_thefts_15_19_pp_f['Num_thefts'] >= 0)], [1])



df_PMX_thefts_15_19_pp_f['pipeline'].sum()

df_PMX_thefts_15_19_pp_f.to_excel('data/df_PMX_thefts_15_19_pp_final.xlsx', index = False)
df_PMX_thefts_15_19_pp_f.to_parquet('data/df_PMX_thefts_15_19_pp_final.parquet')
