==> analysis/08.01_final_merge_data_2019.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_homics_19 = pd.read_parquet("data/df_homics_19.parquet")
df_PMX_19_thefts_pp = pd.read_parquet('data/df_PMX_19_thefts_pp_final.parquet')
df_PMX_gas_place_price = pd.read_parquet('data/df_places_prices_final.parquet')
df_social_2019 = pd.read_excel('data/df_social_19_final_JR.xlsx')
df_SEDENA_19 = pd.read_parquet('data/df_SEDENA_19.parquet')

df_pipeline_total = pd.read_excel('data-raw/df_pipelines.xlsx')


df_cars_2019 = pd.read_parquet('data/df_cars_2019.parquet')
df_rural_w_group = pd.read_parquet('data/df_rural_w_group.parquet')
df_highway_access = pd.read_parquet('data/df_highway_access.parquet')
df_density_km = pd.read_parquet('data/df_density_km.parquet')




df_cvegeo = pd.read_excel("data-raw/codigo municipio inegi.xlsx")
df_cvegeo = df_cvegeo.rename(columns = {"codigo inegi":"CVEGEO"})

df_cvegeo['CVEGEO'] = df_cvegeo['CVEGEO'].astype(str).str.zfill(7).str.replace('.',"")
df_cvegeo['CVEGEO'] = df_cvegeo['CVEGEO'].str[:-1]
df_cvegeo_f = df_cvegeo[["CVEGEO", "NOM_ENT", "NOM_MUN"]]

df_pipeline_total['CVEGEO']=df_pipeline_total['CVEGEO'].astype(object)
df_pipeline_total['CVEGEO'] = df_pipeline_total['CVEGEO'].astype(str).str.zfill(5)



df_social_2019['CVEGEO']=df_social_2019['CVEGEO'].astype(object)
df_social_2019['CVEGEO'] = df_social_2019['CVEGEO'].astype(str).str.zfill(5)

df_PMX_19_thefts_pp

df_cvegeo_f

df_homics_19.dtypes

df_PMX_19_thefts_pp.dtypes

df_pipeline_total

df_PMX_19_thefts_pp_final = pd.merge(df_PMX_19_thefts_pp,
                             df_pipeline_total,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)



dataframes = [df_cvegeo_f,
              df_pipeline_total,
              df_PMX_19_thefts_pp,
              df_PMX_gas_place_price,
              df_homics_19,
              df_SEDENA_19,
              df_social_2019,
              df_cars_2019,
              df_rural_w_group,
              df_highway_access,
              df_density_km]




df_PMX_thefts_19 = reduce(lambda  left,right: pd.merge(left,right,on=['CVEGEO'],
                                            how = 'outer'), dataframes)#.fillna(0)

df_PMX_thefts_19

df_PMX_thefts_19 = df_PMX_thefts_19[(df_PMX_thefts_19['pipe']==1) & (df_PMX_thefts_19['CVEGEO'].str.len()>=5)]

df_PMX_thefts_19["homic_rate"] = df_PMX_thefts_19["homic_19"]/df_PMX_thefts_19["POB_TOT"]*100000
df_PMX_thefts_19["vehic_rate"] = df_PMX_thefts_19["vehicles"]/df_PMX_thefts_19["POB_TOT"]*100000


df_PMX_thefts_19['confront']=df_PMX_thefts_19['confront'].fillna(0)

df_PMX_thefts_19_f = df_PMX_thefts_19.drop(['_merge_x', '_merge_y', "pipeline"], axis=1)


df_PMX_thefts_19_f['pipeline'] = np.select([(df_PMX_thefts_19_f['Num_thefts'] >= 0)], [1])



df_PMX_thefts_19_f['pipeline'].sum()



df_PMX_thefts_19_f

df_PMX_thefts_19_f.to_excel('data/df_PMX_thefts_19_pp_final.xlsx', index = False)
df_PMX_thefts_19_f.to_parquet('data/df_PMX_thefts_19_pp_final.parquet')
