==> analysis/09.01_Final_data_cleaning_2019.ipynb <==
import pandas as pd
import numpy as np


from google.colab import drive
drive.mount('/content/drive')

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_thefts_PMX_complete_ori_2019 = pd.read_excel("data/df_PMX_thefts_19_pp_final.xlsx", converters={'CVEGEO' : str})


df_thefts_PMX_complete_ori_2019.columns

variables_to_maintain = ['CVEGEO','sum_intersects', 'Num_thefts', 
                         'COUNT','price_19', 'POB_TOT', 'confront', 
                         'IRS', 'homic_rate', 'pipeline', 'SBASC', 
                         'POB_6_16_NEB', 'POB_16_EBI', 'NO_SS', 
                         'OVPT', 'NO_SANIT', 'OVSAE', 'NO_DREN', 
                         'NO_ELECT', 'NO_LAVA', 'No_REFRI', "prop_rural",
                         "pob_highway", "porc_highway", "CARR-KM", "LINCARR", "RED-KM",
                        "LINRED", "AREAKM ", "DENSCARR", "DENSRED", "homic_rate", "vehic_rate", "pipeline"]

df_thefts_PMX = df_thefts_PMX_complete_ori_2019[variables_to_maintain].rename( columns = {'CVEGEO': 'Mun_code',
                                 'sum_intersects' : 'Pipe_intersec', 
                                 'Num_thefts' : 'Thefts_cases', 
                                 'COUNT' : 'Gas_stations',
                                 'price_19' : 'Gas_price', 
                                 'POB_TOT' : 'Population', 
                                 'confront' : 'Confront_events',
                                 'IRS' : 'IRS', 
                                 "AREAKM ": "AREAKM",
                                 'homic_rate' : 'Homic_rate', 
                                 'pipeline' : 'Susceptible'})
                                 

df_thefts_PMX

df_thefts_PMX = df_thefts_PMX.fillna(0)
#df_thefts_PMX.fillnan(0)


#df_thefts_PMX[['Gas_price', "Homic_rate"]] = df_thefts_PMX[['Gas_price', "Homic_rate"]].fillna(0)

df_thefts_PMX


df_thefts_PMX['Gas_stations'] = round(df_thefts_PMX['Gas_stations'] / df_thefts_PMX['Population'] * 100000, 2)
df_thefts_PMX['Gas_price'] = round(df_thefts_PMX['Gas_price'],2)
df_thefts_PMX['Homic_rate'] = round(df_thefts_PMX['Homic_rate'],2)
df_thefts_PMX['IRS'] = round(df_thefts_PMX['IRS'],4)
df_thefts_PMX['Thefts_cases'] = df_thefts_PMX['Thefts_cases'].fillna(0)
df_thefts_PMX_cleaned = df_thefts_PMX.drop(['Population', 'Thefts_cases'], axis = 1, inplace = False)
df_thefts_PMX_cleaned

df_thefts_PMX_cleaned.head()


df_thefts_PMX_cleaned.to_excel('data/df_thefts_PMX_final_2019.xlsx', index = False)
